Separate sort-group-filter into separate functions.
